Improvement modeling to suport bigdecimal to price and

+++ src/main/java/com/restaurant/searcher/domain/model/rest/response/RestaurantRestResponse.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
@@ -12,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class RestaurantRestResponse {
     private String name;
     private Integer customerRating;
-    private Integer distance;
-    private Integer price;
+    private BigDecimal distance;
+    private BigDecimal price;
     private String cuisineName;
 }
